setup for culster graph
