+++ frontend/src/results/d3/index.js
@@ -1,4 +1,4 @@
 // Exports from this module
 export { default as CompareUsers } from './compareUsers.js'
-export { default as D3BarChart } from './d3BarChart.js'
 export { default as GenreStatistics } from './genreStatistics.js'
+export { default as ClusterUsers } from './cultserUsers.js'
